Move resource actions to espace producteur controller
