+++ apps/transport/lib/transport_web/templates/espace_producteur/edit_dataset.html.heex
@@ -49,13 +49,13 @@
               <td><%= resource["title"] %> <span :if={format} class="label"><%= format %></span></td>
               <td class="align-right">
                 <%= link(dgettext("espace-producteurs", "Edit resource"),
-                  to: resource_path(@conn, :form, @datagouv_dataset["id"], resource["id"]),
+                  to: espace_producteur_path(@conn, :resource_form, @datagouv_dataset["id"], resource["id"]),
                   class: "button-outline primary small",
                   "data-tracking-category": "espace_producteur",
                   "data-tracking-action": "update_resource_button"
                 ) %>
                 <%= link(dgettext("espace-producteurs", "Delete resource"),
-                  to: resource_path(@conn, :delete_resource_confirmation, @datagouv_dataset["id"], resource["id"]),
+                  to: espace_producteur_path(@conn, :delete_resource_confirmation, @datagouv_dataset["id"], resource["id"]),
                   class: "button-outline warning small",
                   "data-tracking-category": "espace_producteur",
                   "data-tracking-action": "delete_resource_button"
@@ -66,7 +66,7 @@
         </table>
 
         <%= link(dgettext("espace-producteurs", "Add a resource"),
-          to: resource_path(@conn, :form, @dataset.datagouv_id),
+          to: espace_producteur_path(@conn, :resource_form, @dataset.datagouv_id),
           class: "button primary",
           "data-tracking-category": "espace_producteur",
           "data-tracking-action": "add_resource_button"
