+++ apps/transport/lib/transport_web/views/espace_producteur_view.ex
@@ -1,4 +1,32 @@
 defmodule TransportWeb.EspaceProducteurView do
   use TransportWeb, :view
   import TransportWeb.BreadCrumbs, only: [breadcrumbs: 1]
+
+
+  @spec action_path(Plug.Conn.t()) :: any
+  def action_path(%Plug.Conn{params: %{"resource_id" => r_id} = params} = conn),
+    do: espace_producteur_path(conn, :post_file, params["dataset_id"], r_id)
+
+  def action_path(%Plug.Conn{params: params} = conn),
+    do: espace_producteur_path(conn, :post_file, params["dataset_id"])
+
+  def title(%Plug.Conn{params: %{"resource_id" => _}}),
+    do: dgettext("resource", "Resource modification")
+
+  def title(_), do: dgettext("resource", "Add a new resource")
+
+  def remote?(%{"filetype" => "remote"}), do: true
+  def remote?(_), do: false
+
+  def link_to_datagouv_resource_edit(dataset_id, resource_id),
+    do:
+      :transport
+      |> Application.fetch_env!(:datagouvfr_site)
+      |> Path.join("/fr/admin/dataset/#{dataset_id}/resource/#{resource_id}")
+
+  def link_to_datagouv_resource_creation(dataset_id),
+    do:
+      :transport
+      |> Application.fetch_env!(:datagouvfr_site)
+      |> Path.join("/fr/admin/dataset/#{dataset_id}?new_resource=")
 end
